done with complete CRUD

+++ base/views.py
@@ -2,7 +2,7 @@ from django import forms
 from django.http.response import HttpResponseRedirect
 from django.shortcuts import render,redirect
 from .models import Message, Room, Topic
-from .forms import RoomForm
+from .forms import MessageForm, RoomForm
 from django.db.models import Q
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import  User
@@ -67,17 +67,31 @@ def home(request):
         Q(host__username__icontains=q)
         
         )
+    
     topics = Topic.objects.all()
     room_count = rooms.count()
-    context ={'rooms':rooms,  'topics': topics, 'room_count' : room_count}
+    room_message = Message.objects.filter(Q(room__topic__name__icontains=q))
+    context ={'rooms':rooms,  'topics': topics, 'room_count' : room_count, 'room_message':room_message }
     return render (request, 'chat/index.html',context)
 
 
 
+def ProfilePage(request,pk):
+    users = User.objects.get(id=pk)
+    rooms =users.room_set.all()
+    room_msg =users.message_set.all()
+    print(room)
+    context ={'users':users, 'rooms':rooms, 'room_msg': room_msg}
+    return render(request, 'chat/profile.html', context)
+
+
+
+# room with ID function
 
 def room(request, pk):
     room = Room.objects.get(id =pk)
-    room_messages = room.message_set.all().order_by('-created') #_set.all() is used for One-to-many relationship
+    referrer = request.META['HTTP_REFERER']
+    room_messages = room.message_set.all() #_set.all() is used for One-to-many relationship
     participants = room.participants.all() #.all() is used for Many-To-Many relationship
     if request.method == 'POST':
         message = Message.objects.create(
@@ -137,16 +151,52 @@ def DeleteRoom(request,pk):
         messages.error(request, 'Sorry, You are not permited to delete this room')
         print(referrer)
         return HttpResponseRedirect(referrer)
-    if request.method == 'POST':
-        room.delete()
-        return redirect ('home')
+    if request.User.is_superuser == True:
+        if request.method == 'POST':
+            room.delete()
+            return redirect ('home')
     return render (request, 'chat/delete.html', {'obj': room})
 
 
+# Delete message
+@login_required(login_url='/login')
+def DeleteMessage(request,pk):
+    room_message = Message.objects.get(id=pk)
+    if request.method == 'POST':
+         if request.user == room_message.user or request.user.is_superuser == True:
+                print(request.user)
+                room_message.delete()
+                messages.success(request, 'Message deleted successfully')
+                print(request.user.is_superuser)
+                return redirect ('home')
+         else:
+             referrer = request.META['HTTP_REFERER']
+             messages.error(request, 'Sorry, You are not permited to delete this message')
+             print(referrer)
+             return HttpResponseRedirect(referrer)
+    return render (request, 'chat/delete.html', {'obj': room_message})
 
 
 
 
+
+# Update room message here
+@login_required(login_url='/login')
+def UpdateMessage(request, pk):
+    msg_update =Message.objects.get(id=pk)
+    form =MessageForm(instance=msg_update)
+    if request.user != msg_update.user:
+         referrer = request.META['HTTP_REFERER']
+         messages.error(request, 'Sorry, You are not permited to updated this message')
+         print(referrer)
+         return HttpResponseRedirect(referrer)
+    if request.method == 'POST':
+        form = MessageForm(request.POST, instance=msg_update)
+        if form.is_valid(): 
+            form.save()
+            return redirect('home')
+    context ={'form':form}
+    return render(request, 'chat/room_form.html', context)
 
 
 # rooms = [
